prevent overflow

+++ Assets/Scripts/Manager/SceneIntro.cs
@@ -6,7 +6,7 @@ using Random = UnityEngine.Random;
 public class SceneIntro : MonoBehaviour{
     private void Update() {
         if (Input.anyKeyDown) {
-            var transition = (Trans) ((int) Random.Range(1, 0b10000)) + ((int) Random.Range(0b100000, 0b1000000000));
+            var transition = (Trans) (((long) Random.Range(1, 0b10000)) + ((long) Random.Range(0b100000, 0b1000000000)));
             // transition = Trans.FromRight | Trans.ToLeft 등으로 뱡향 설정
             GameManager.Instance.Transition(transition, () => Debug.Log("asdf"));
         }
